Extract tag functionality of TriggerOnTagEnter to inheritance parent
